Moved the progress bar file into the context module, now that this is effectively a member of the context struct.
